Make the functions to remove contacts split their workload in batches
If e_book_async_remove_contacts() or e_book_remove_contacts() were
called on a lot of contacts then it was possible to get a D-Bus timeout
as the backend couldn't complete the operation in less than 25 seconds.
Now we make a D-Bus call for at most BATCH_THRESHOLD contacts.

Fixes: NB#138862 - Addressbook get crashed while deleting all the
contacts