Make sure to stop the view once and only once
priv->running must be set to FALSE *before* stopping the view. This is because
stopping the view could dispose it and the dispose function will try to stop it
again.

Also document and clean a bit the live cycle of EDataBook/EDataBookView. It is
really a mess.

Fixes NB#160630 - e-addressbook-factory new crash trace in slice_alloc after fix of FD & mem leak